recipes-tisdk: Support LDC colorconversion
As our cameras natively output interleaved formats and do not support
planar formats such as NV12, we required a more efficient way to perform
colorspace conversion than to use the DSPs. Using the LDC, which has
built-in capability for color conversion, we can easily do this.

Note, at this point, when the LDC is patched this way, it will only
serve as a color conversion block. Future modifications will be needed
in order for the LDC to serve to perform lens distortion correction as
well as color conversion