recipes-core:packagegroups: Remove native PRU Compiler
A native PRU toolchain is included in some tisdk images. However
this causes build problems on ARM64 hosts (e.g. Apple M-cores) as
there is a dependency on a x86 toolchain.  As we do not have
plans to support the PRU on ROVY products, we can simply remove
this for now.