recipes-core:images:tisdk-edgeai-image: Add bbappend
The default tisdk-edgeai-image does not install tisdk-uenv, which provides
uEnv.txt. The Arago distro requires uEnv.txt, but as part of IMAGE_BOOT_FILES
in the wic image (in the boot partition).  This bbappend adds the tisdk-uenv
package to IMAGE_INSTALL so that uEnv.txt is installed to /boot/uEnv.txt.